requirements.txt
rasterio>=1.3.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
click>=8.1.0
tqdm>=4.65.0
